define new make_default_test_cluster and remove old implementations
