+++ tests/test_utils/test_api_common.hpp
@@ -94,20 +94,6 @@ inline bool is_galaxy_configuration(Cluster* cluster) {
             cluster->get_cluster_description()->get_board_type(0) == tt::BoardType::UBB_BLACKHOLE);
 }
 
-inline bool has_remote_chips() {
-    std::vector<int> pci_device_ids = PCIDevice::enumerate_devices();
-    if (pci_device_ids.empty()) {
-        return false;
-    }
-    std::unique_ptr<TTDevice> tt_device = TTDevice::create(pci_device_ids[0]);
-    tt_device->init_tt_device();
-
-    auto board_type = tt_device->get_board_type();
-    return board_type == tt::BoardType::N300;
-}
-
-inline uint32_t get_num_host_ch_for_test() { return has_remote_chips() ? 1UL : 0UL; }
-
 // Returns the top-left (lowest x, lowest y) and bottom-right (highest x, highest y) TENSIX cores
 // in translated coordinates for the given SoC descriptor.
 inline std::vector<CoreCoord> get_tensix_corners(const SocDescriptor& soc_desc) {
@@ -138,15 +124,3 @@ constexpr uint64_t SAFE_IO_L1_ADDRESS = 0x1000;
 // True when the test should run against a simulator, indicated by TT_UMD_SIMULATOR
 // pointing at the simulator binary directory.
 inline bool is_simulation_test() { return std::getenv("TT_UMD_SIMULATOR") != nullptr; }
-
-// Creates a Cluster for an API test. If TT_UMD_SIMULATOR is set the chip_type,
-// target_devices, and simulator_directory fields of `options` are overridden to
-// target the simulator; otherwise `options` is used as-is (default = silicon).
-inline std::unique_ptr<Cluster> make_cluster_for_test(ClusterOptions options = {}) {
-    if (const char* sim_path = std::getenv("TT_UMD_SIMULATOR")) {
-        options.chip_type = ChipType::SIMULATION;
-        options.target_devices = {0};
-        options.simulator_directory = std::filesystem::path(sim_path);
-    }
-    return std::make_unique<Cluster>(options);
-}
